@@ -15,22 +15,22 @@
   "dependencies": {
     "@types/core-js": "^2.5.0",
     "@types/keymaster": "^1.6.28",
-    "@types/lodash": "^4.14.129",
-    "@types/react": "^16.8",
-    "@types/react-dom": "^16.8",
+    "@types/lodash": "^4.14.151",
+    "@types/react": "=16.8.2",
+    "@types/react-dom": "=16.8.0",
     "@webscopeio/react-textarea-autocomplete": "=1.4.1",
     "blacklist": "^1.1.4",
     "bootstrap": "3.1.1",
-    "classnames": "^2.2.5",
-    "create-react-class": "^15.6.2",
+    "classnames": "^2.2.6",
+    "create-react-class": "^15.6.3",
     "eventemitter3": "^3.0.1",
     "fileapi": "^2.0.21",
     "keymaster": "^1.6.2",
     "markdown-it": "^8.4.2",
     "moment": "^2.24.0",
-    "react": "^16.8",
+    "react": "^16.13.1",
     "react-bootstrap": "^0.32.3",
-    "react-dom": "^16.8",
+    "react-dom": "^16.13.1",
     "react-dom-factories": "^1.0.2",
     "react-router-dom": "^4.3.1",
     "react-select": "1.0.0-rc.10",
@@ -39,7 +39,7 @@
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@types/assert": "^1.4.6",
+    "@types/assert": "^1.4.7",
     "@types/minimist": "^1.2.0",
     "@types/node": "^14.0.1",
     "@wdio/cli": "^6.1.11",
