@@ -42,7 +42,7 @@ export const UserNotifications = createFactory({
   },
 
   // SHOULD Switch to componentDidUpdate instead, see  users-page.more.ts  for how.
-  componentWillReceiveProps: function(nextProps: any) {
+  UNSAFE_componentWillReceiveProps: function(nextProps: any) {
     // Dupl code, also in view drafts. [7WUBKZ0]
     const me: Myself = this.props.store.me;
     const user: UserInclDetails = this.props.user;
