@@ -28,8 +28,8 @@ export function createComponent(componentDefinition): any { // oops should obvio
     // The mere presence of these functions cause an unknown error when rendering
     // React-Router server side. So remove them; they're never called server side anyway.
     // The error logs the message '{}' to console.error(); no idea what that means.
-    delete componentDefinition.componentWillUpdate;
-    delete componentDefinition.componentWillReceiveProps;
+    delete componentDefinition.UNSAFE_componentWillUpdate;
+    delete componentDefinition.UNSAFE_componentWillReceiveProps;
   }
   return reactCreateFactory(createReactClass(componentDefinition));
 }
